DRTII-762 Further refactor to eliminate old `tocsv` objects
